requirements.txt
aiogram>=3.3.0,<4.0.0
google-genai>=1.57.0
psycopg2-binary==2.9.9